Consolidate shared code
